requirements.txt
aiosignal==1.3.1
attrs==23.1.0
boto3==1.28.59
botocore==1.31.59
certifi==2023.7.22
charset-normalizer==3.2.0
click==8.1.7
confluent-kafka==2.3.0
dnspython==2.4.2
exceptiongroup==1.1.3
filelock==3.12.4
frozenlist==1.4.0
idna==3.4
iniconfig==2.0.0
jmespath==1.0.1
jsonschema==4.19.1
jsonschema-specifications==2023.7.1
kafka-python==2.0.2
msgpack==1.0.7
numpy==1.26.1
packaging==23.2
pluggy==1.3.0
protobuf==4.24.4
py4j==0.10.9.7
pymongo==4.5.0
pyspark==3.5.0
pytest==7.4.2
python-dateutil==2.8.2
PyYAML==6.0.1
ray==2.7.1
referencing==0.30.2
requests==2.31.0
retrying==1.3.4
rpds-py==0.10.6
s3transfer==0.7.0
six==1.16.0
tomli==2.0.1
urllib3==1.26.17
